:recycle: move logic into custom component
